pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "axiom-trace"
version = "1.1.0"
description = "Local-first, append-only trace vault for AI agents with cryptographic integrity"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Axiom Team" }
]
keywords = ["ai", "tracing", "accountability", "vector-search", "memvid"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "memvid==0.1.3",
    "typer==0.12.3",
    "jsonschema==4.22.0",
    "orjson==3.10.7",
    "filelock>=3.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
axiom = "axiom_trace.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["axiom_trace"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
